requirements.txt
Django>=3.2
djangorestframework==3.14.0
openpyxl==3.1.2
orjson==3.9.15
pandas==2.2.1
psycopg2==2.9.9
python-dateutil==2.8.2
sqlparse==0.4.4
ujson==5.9.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
